+++ packages/gitmoji-changelog-cli/src/cli.js
@@ -1,10 +1,14 @@
-const { generateChangelog } = require('@gitmoji-changelog/core')
+const { generateChangelog, logger } = require('@gitmoji-changelog/core')
 const { convert } = require('@gitmoji-changelog/markdown')
 const fs = require('fs')
 
 async function main({ format } = {}) {
   try {
     const changelog = await generateChangelog()
+
+    logger.info(changelog.meta.package.name)
+    logger.info('v%s', changelog.meta.package.version)
+    logger.info(changelog.meta.repository.url)
 
     switch (format) {
       case 'json':